Data-Engineering/1-Extract-BM/requirements.txt
google-cloud-storage
pandas
wbgapi